+++ src/main/java/com/koreait/dooboo/communityboard/controller/CommunityBoardController.java
@@ -1,56 +1,39 @@
 package com.koreait.dooboo.communityboard.controller;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-
-import javax.annotation.Resource;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.koreait.dooboo.communityboard.command.BoardCommand;
-import com.koreait.dooboo.communityboard.command.BoardCommandImpl;
 import com.koreait.dooboo.communityboard.dto.CommunityBoardDTO;
 
 @Controller
-@RequestMapping("/CommunityBoard")
+/*@RequestMapping("/CommunityBoard")*/
 public class CommunityBoardController {
 	
 	@Autowired
 	private BoardCommand boardCommand;
 	
 	
-	@GetMapping(value = "/board_sample")
+	@GetMapping(value = "b.list")
 	public ModelAndView openBoardList(CommunityBoardDTO boardDTO) {
 		ModelAndView mav = new ModelAndView();
 		
 		Map<String,Object> resultMap = boardCommand.getBoardList(boardDTO);
 		mav.addObject("resultMap", resultMap);
 		mav.addObject("params", boardDTO);
-		mav.setViewName("CommunityBoard/board_sample");
+		mav.setViewName("CommunityBoard/boardList");
 		return mav;
 	}
 	
-	@GetMapping(value = "/detail")
+	@GetMapping(value = "b.detail")
 	public ModelAndView openBoardDetail(CommunityBoardDTO boardDTO) {
-		
-		ModelAndView mav = new ModelAndView();
-		CommunityBoardDTO resultMap = boardCommand.getBoardDetail(boardDTO.getBoardNo());
-		
-		mav.addObject("board", resultMap);
-		mav.setViewName("CommunityBoard/write");
-		
-		return mav; 
-	}
-	
-	@GetMapping(value = "/read")
-	public ModelAndView openBoardRead(CommunityBoardDTO boardDTO) {
 		
 		ModelAndView mav = new ModelAndView();
 		CommunityBoardDTO resultMap = boardCommand.getBoardDetail(boardDTO.getBoardNo());
@@ -58,11 +41,23 @@ public class CommunityBoardController {
 		mav.addObject("board", resultMap);
 		mav.setViewName("CommunityBoard/read");
 		
+		return mav; 
+	}
+	
+	@GetMapping(value = "b.write")
+	public ModelAndView openBoardRead(CommunityBoardDTO boardDTO) {
+		
+		ModelAndView mav = new ModelAndView();
+		CommunityBoardDTO resultMap = boardCommand.getBoardDetail(boardDTO.getBoardNo());
+		
+		mav.addObject("board", resultMap);
+		mav.setViewName("CommunityBoard/write");
+		
 		return mav;
 	}
 	
 	
-	@GetMapping(value = "/update")
+	@GetMapping(value = "b.update")
 	public ModelAndView openBoardUpdate(CommunityBoardDTO boardDTO) {
 		
 		ModelAndView mav = new ModelAndView();
@@ -74,7 +69,7 @@ public class CommunityBoardController {
 		return mav;
 	}
 	 
-	@PostMapping(value ="/save")
+	@PostMapping(value ="b.save")
 	@ResponseBody
 	public Map<String,Object> openBoardSave(CommunityBoardDTO boardDTO) {
 		boolean saveYn = boardCommand.registerBoard(boardDTO);
@@ -83,10 +78,10 @@ public class CommunityBoardController {
 		return resultMap;
 	}
 
-	@PostMapping(value ="/delete")
+	@PostMapping(value ="b.delete")
 	@ResponseBody
 	public Map<String,Object> openBoardDelete(CommunityBoardDTO boardDTO) {
-		boolean deleteYn = boardCommand.deleteBoard(boardDTO.getBoardNo());
+		boolean deleteYn = boardCommand.deleteBoard((int) boardDTO.getBoardNo());
 		Map<String,Object> resultMap = new HashMap<String,Object>();
 		resultMap.put("deleteYn", deleteYn);
 		return resultMap;
