+++ src/main/webapp/WEB-INF/views/CommunityBoard/boardList.jsp
@@ -13,7 +13,7 @@
 </header>
 <body>
 	<div>
-          <table>
+        <table style="border: solid 1px;width : 1000px;">
               <thead>
                  <tr>
 		            <th>번호</th>
@@ -26,35 +26,35 @@
                 </tr>
               </thead>
                <tbody>
-                 	<c:if test="${empty resultMap.boardList}">
-						<tr>
-							<th colspan="7">검색 결과가 없습니다.</th>
-						</tr>
-					</c:if>
-					<c:if test="${not empty resultMap.boardList}">
-					        <c:forEach items="${resultMap.boardList}" var="post">
-				            	<tr>
-					                <td>${post.getRownum()}</td>
-					                <td><a href="/CommunityBoard/read?boardNo=${post.getBoardNo()}">${post.getTitle()}</a></td>
-					                <td>${post.getNickname()}</td>
-					                <td>${post.getRegDate()}</td>
-					                <td>${post.getModDate()}</td>
-					                <td>${post.getHit()}</td>
-					                <td>
-					                	<a href="/CommunityBoard/update?boardNo=${post.getBoardNo()}" style="color: green;">수정</a><br>
-					                	<a href="javascript:boardDelete(${post.getBoardNo()})" style="color: green;">삭제</a>
-					                </td>
-	          				  	</tr>
-	       				 </c:forEach>
-					</c:if>
+	                 	<c:if test="${empty resultMap.boardList}">
+							<tr>
+								<th colspan="7">검색 결과가 없습니다.</th>
+							</tr>
+						</c:if>
+						<c:if test="${not empty resultMap.boardList}">
+						        <c:forEach items="${resultMap.boardList}" var="post">
+					            	<tr>
+						                <td>${post.getRownum()}</td>
+						                <td><a href="b.detail?boardNo=${post.getBoardNo()}">${post.getTitle()}</a></td>
+						                <td>${post.getNickname()}</td>
+						                <td>${post.getRegDate()}</td>
+						                <td>${post.getModDate()}</td>
+						                <td>${post.getHit()}</td>
+						                <td>
+						                	<a href="b.update?boardNo=${post.getBoardNo()}" style="color: green;">수정</a><br>
+						                	<a href="javascript:boardDelete(${post.getBoardNo()})" style="color: green;">삭제</a>
+						                </td>
+		          				  	</tr>
+		       				 </c:forEach>
+						</c:if>
+			        </table>
 			</tbody>
-   		</table>
  	</div>
 		<div>
 			<ul>	
-	  		<%-- <if test="${resultMap.listPagingInfo.hasPreviousPage }">
+	  		<if test="${resultMap.listPagingInfo.hasPreviousPage }">
 			    <li>
-				      <a href="/board/list?currentPageNo=${status.index +1}"  tabindex="-1" aria-label="Previous">
+				      <a href="b.list?currentPageNo=${status.index +1}"  tabindex="-1" aria-label="Previous">
 					        <span aria-hidden="true">«</span>
 					        <span>Previous</span>
 				      </a>
@@ -63,24 +63,22 @@
 		    <c:forEach begin="${resultMap.listPagingInfo.firstPage}" end="${resultMap.listPagingInfo.lastPage}" varStatus="status">
 		    <li>  
 		    	<c:if test="${param.currentPageNo eq  status.index}">active</c:if>">
-			      <a href="/CommunityBoard/list?currentPageNo=${status.index}">${status.index}
-		        	<span>(current)</span>
-			      </a>
+			      <a href="b.list?currentPageNo=${status.index}">${status.index}</a>
 		    </li>
 		  	</c:forEach>
 		  	<if test="${resultMap.listPagingInfo.hasNextPage}">
 			    <li>
-			      	<a href="/CommunityBoard/list?currentPageNo=${resultMap.listPagingInfo.lastPage+1 }"aria-label="Next">
+			      	<a href="b.list?currentPageNo=${resultMap.listPagingInfo.lastPage+1 }"aria-label="Next">
 			        	<span aria-hidden="true">»</span>
 			        	<span>Next</span>
 			      </a>
 			    </li>
 			 </if>  
-			 <button><a href="/CommunityBoard/update">글쓰기</a></button>		
-	  </ul> --%>
+			 <button><a href="b.write">글쓰기</a></button>		
+	  </ul>
 	</div>
-		<!-- 
-		<form id="search"  action="/CommunityBoard/list" method="get" style="animation-duration: 300ms; right: 0px;" >
+		
+		<form id="search"  action="b.list" method="get" style="animation-duration: 300ms; right: 0px;" >
 			<div>
 		  		<select  name="searchType" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"  style="width:150px;">
 					    <option value="">전체</option>
@@ -94,7 +92,7 @@
                      </div>
                   </div>
 			  </div>
-		 </form> -->
+		 </form>
 	
 <script type="text/javascript">
 	var searchForm = $("search");
@@ -113,7 +111,7 @@
 	function boardDelete(boardNo){
 		if(confirm('삭제 하시겠습니까?')){
 			$.ajax({
-			    url: "/board/delete", // 클라이언트가 요청을 보낼 서버의 URL 주소
+			    url: "/dooboo/b.delete", // 클라이언트가 요청을 보낼 서버의 URL 주소
 			    data: {
 			    	'boardNo' : boardNo
 			    },                // HTTP 요청과 함께 서버로 보낼 데이터
@@ -122,7 +120,7 @@
 		        success : function(data) {
 		        	console.log(data)
 		        	alert('삭제 되었습니다.');
-		       		location.href='/board/list'
+		       		location.href='/dooboo/b.list'
 		        }, // success 
 		        error : function(xhr, status) {
 		        	alert('error')
